requirements.txt
Flask
opencv-python
numpy
gunicorn
pyheif
Pillow
